将 ff14 插件拆成多个子插件，并适配元信息
