+++ src/plugins/ff14/config.py
@@ -2,7 +2,6 @@
 """
 from typing import List
 
-from nonebot import get_driver
 from nonebot_plugin_datastore import PluginData
 from pydantic import BaseSettings, validator
 
@@ -44,7 +43,3 @@ class Config(BaseSettings):
     class Config:
         extra = "ignore"
         validate_assignment = True
-
-
-global_config = get_driver().config
-plugin_config = Config.parse_obj(global_config)
