+++ src/plugins/ff14/plugins/fflogs/fflogs_data.py
@@ -6,7 +6,7 @@ from typing import Optional
 
 from pydantic import BaseModel
 
-from .config import DATA
+from ... import DATA
 
 
 class BossInfo(BaseModel):
