+++ tests/plugins/ff14/test_gate.py
@@ -6,27 +6,25 @@ from tests.fake import fake_group_message_event
 
 
 @pytest.mark.asyncio
-@pytest.mark.parametrize("app", [("src.plugins.ff14",)], indirect=True)
-async def test_gete(
-    app: App,
-    mocker: MockerFixture,
-):
+async def test_gete(app: App, mocker: MockerFixture):
     """测试藏宝选门，两个门的情况"""
-    from nonebot import get_driver
-    from nonebot.adapters.onebot.v11 import Adapter, Bot, Message
-    from nonebug.mixin.call_api.fake import make_fake_adapter, make_fake_bot
+    from nonebot import require
+    from nonebot.adapters.onebot.v11 import Message
 
-    from src.plugins.ff14 import gate_cmd
-    from src.plugins.ff14.gate import EXPR_GATE
+    require("src.plugins.ff14")
 
-    render_expression = mocker.patch("src.plugins.ff14.gate.render_expression")
+    from src.plugins.ff14.plugins.gate import gate_cmd
+    from src.plugins.ff14.plugins.gate.data_source import EXPR_GATE
+
+    render_expression = mocker.patch(
+        "src.plugins.ff14.plugins.gate.data_source.render_expression"
+    )
     render_expression.return_value = Message("test")
-    randint = mocker.patch("src.plugins.ff14.gate.randint")
+    randint = mocker.patch("src.plugins.ff14.plugins.gate.data_source.randint")
     randint.return_value = 1
 
     async with app.test_matcher(gate_cmd) as ctx:
-        adapter = make_fake_adapter(Adapter)(get_driver(), ctx)
-        bot = make_fake_bot(Bot)(adapter, "1")
+        bot = ctx.create_bot()
         event = fake_group_message_event(message=Message("/gate 2"))
 
         ctx.receive_event(bot, event)
@@ -38,27 +36,25 @@ async def test_gete(
 
 
 @pytest.mark.asyncio
-@pytest.mark.parametrize("app", [("src.plugins.ff14",)], indirect=True)
-async def test_gete_ask_arg(
-    app: App,
-    mocker: MockerFixture,
-):
+async def test_gete_ask_arg(app: App, mocker: MockerFixture):
     """测试藏宝选门，两个门的情况"""
-    from nonebot import get_driver
-    from nonebot.adapters.onebot.v11 import Adapter, Bot, Message
-    from nonebug.mixin.call_api.fake import make_fake_adapter, make_fake_bot
+    from nonebot import require
+    from nonebot.adapters.onebot.v11 import Message
 
-    from src.plugins.ff14 import gate_cmd
-    from src.plugins.ff14.gate import EXPR_GATE
+    require("src.plugins.ff14")
 
-    render_expression = mocker.patch("src.plugins.ff14.gate.render_expression")
+    from src.plugins.ff14.plugins.gate import gate_cmd
+    from src.plugins.ff14.plugins.gate.data_source import EXPR_GATE
+
+    render_expression = mocker.patch(
+        "src.plugins.ff14.plugins.gate.data_source.render_expression"
+    )
     render_expression.return_value = Message("test")
-    randint = mocker.patch("src.plugins.ff14.gate.randint")
+    randint = mocker.patch("src.plugins.ff14.plugins.gate.data_source.randint")
     randint.return_value = 1
 
     async with app.test_matcher(gate_cmd) as ctx:
-        adapter = make_fake_adapter(Adapter)(get_driver(), ctx)
-        bot = make_fake_bot(Bot)(adapter, "1")
+        bot = ctx.create_bot()
         event = fake_group_message_event(message=Message("/gate"))
         next_event = fake_group_message_event(message=Message("2"))
 
@@ -73,27 +69,25 @@ async def test_gete_ask_arg(
 
 
 @pytest.mark.asyncio
-@pytest.mark.parametrize("app", [("src.plugins.ff14",)], indirect=True)
-async def test_gete_ask_arg_error(
-    app: App,
-    mocker: MockerFixture,
-):
+async def test_gete_ask_arg_error(app: App, mocker: MockerFixture):
     """测试藏宝选门，两个门的情况，第一次输入错误"""
-    from nonebot import get_driver
-    from nonebot.adapters.onebot.v11 import Adapter, Bot, Message
-    from nonebug.mixin.call_api.fake import make_fake_adapter, make_fake_bot
+    from nonebot import require
+    from nonebot.adapters.onebot.v11 import Message
 
-    from src.plugins.ff14 import gate_cmd
-    from src.plugins.ff14.gate import EXPR_GATE
+    require("src.plugins.ff14")
 
-    render_expression = mocker.patch("src.plugins.ff14.gate.render_expression")
+    from src.plugins.ff14.plugins.gate import gate_cmd
+    from src.plugins.ff14.plugins.gate.data_source import EXPR_GATE
+
+    render_expression = mocker.patch(
+        "src.plugins.ff14.plugins.gate.data_source.render_expression"
+    )
     render_expression.return_value = Message("test")
-    randint = mocker.patch("src.plugins.ff14.gate.randint")
+    randint = mocker.patch("src.plugins.ff14.plugins.gate.data_source.randint")
     randint.return_value = 1
 
     async with app.test_matcher(gate_cmd) as ctx:
-        adapter = make_fake_adapter(Adapter)(get_driver(), ctx)
-        bot = make_fake_bot(Bot)(adapter, "1")
+        bot = ctx.create_bot()
         event = fake_group_message_event(message=Message("/gate"))
         next_event = fake_group_message_event(message=Message("4"))
         final_event = fake_group_message_event(message=Message("2"))
